+++ src/DZ_3/Dz_3_2/ExcelentPupil.java
@@ -1,0 +1,22 @@
+package DZ_3.Dz_3_2;
+
+public class ExcelentPupil extends Pupil{
+    @Override
+    public void study() {
+        System.out.println("Excelent study");
+    }
+    @Override
+    public void write() {
+        System.out.println("Excelent write");
+    }
+
+    @Override
+    public void read() {
+        System.out.println("Excelent read");
+    }
+
+    @Override
+    public void relax() {
+        System.out.println("Excelent relax");
+    }
+}
