+++ src/DZ_3/Dz_3_2/GoodPupil.java
@@ -1,0 +1,24 @@
+package DZ_3.Dz_3_2;
+
+public class GoodPupil extends  Pupil{
+    @Override
+    public void study() {
+        System.out.println("Good study");
+    }
+
+    @Override
+    public void write() {
+        System.out.println("Good write");
+    }
+
+    @Override
+    public void read() {
+        System.out.println("Good read");
+    }
+
+    @Override
+    public void relax() {
+        System.out.println("Good relax");
+    }
+
+}
