+++ src/DZ_3/Dz_3_2/KlassRoom.java
@@ -1,0 +1,13 @@
+package DZ_3.Dz_3_2;
+
+public class KlassRoom {
+    Pupil pupil = new Pupil();
+
+    public KlassRoom() {
+        pupil.study();
+        pupil.read();
+        pupil.write();
+        pupil.relax();
+    }
+
+}
