+++ src/DZ_3/Dz_3_2/Pupil.java
@@ -1,0 +1,16 @@
+package DZ_3.Dz_3_2;
+
+public class Pupil {
+
+    public void study (){
+    }
+    public void read(){
+
+    }
+    public void write(){
+
+    }
+    public  void relax(){
+
+    }
+}
