+++ src/DZ_3/Dz_3_3/Vehicle.java
@@ -1,0 +1,15 @@
+package DZ_3.Dz_3_3;
+
+public class Vehicle {
+    public int coordinates;
+    public int prices;
+    public int speed;
+    public  int ears;
+
+    public Vehicle(int coordinates, int prices, int speed, int ears) {
+        this.coordinates = coordinates;
+        this.prices = prices;
+        this.speed = speed;
+        this.ears = ears;
+    }
+}
