+++ src/DZ_3/Dz_3_4/ExpertDocumentWorker.java
@@ -1,0 +1,10 @@
+package DZ_3.Dz_3_4;
+
+public class ExpertDocumentWorker extends ProDocumentWorker {
+        final int exp_kay=6492;
+    @Override
+    public void saveDocument() {
+        System.out.println("Документ сохранен в новом формате");
+    }
+
+}
