+++ src/DZ_4/Dz_4_2/AbstractHandler.java
@@ -1,0 +1,16 @@
+package DZ_4.Dz_4_2;
+
+abstract class AbstractHandler {
+    void open(){
+
+    }
+    void create(){
+
+    }
+    void change(){
+
+    }
+    void  save(){
+
+    }
+}
