+++ src/DZ_4/Dz_4_3/Playable.java
@@ -1,0 +1,7 @@
+package DZ_4.Dz_4_3;
+
+public interface Playable {
+     void  play();
+     void  pause();
+     void record();
+}
