+++ src/DZ_4/Dz_4_3/Player.java
@@ -1,0 +1,33 @@
+package DZ_4.Dz_4_3;
+
+public class Player implements Playable,Recodable{
+    @Override
+    public void play() {
+
+    }
+
+    @Override
+    public void pause() {
+
+    }
+
+    @Override
+    public void record() {
+
+    }
+
+    @Override
+    public void recoer() {
+
+    }
+
+    @Override
+    public void paose() {
+
+    }
+
+    @Override
+    public void stop() {
+
+    }
+}
