+++ src/DZ_4/Dz_4_3/Recodable.java
@@ -1,0 +1,7 @@
+package DZ_4.Dz_4_3;
+
+public interface Recodable {
+    void  recoer();
+    void paose();
+    void stop();
+}
