+++ src/Dz_2/Dz_2_2/Car.java
@@ -1,0 +1,16 @@
+package Dz_2.Dz_2_2;
+
+public class Car {
+    public int year;
+    public String color;
+
+    public Car(int year) {
+        this.year = year;
+    }
+
+    public Car(int year, String color) {
+        this.year = year;
+        this.color = color;
+    }
+
+}
