--- src/Dz_2_2/Car_main.java
@@ -1,10 +1,0 @@
-package Dz_2_2;
-
-public class Car_main {
-    public static void main(String[] args) {
-        Dz_2_2.Car car1 = new Dz_2_2.Car(45);
-        Dz_2_2.Car car2 = new Dz_2_2.Car(64,"Black");
-        System.out.println(car1.year);
-        System.out.println(car2.year +"\t"+ car2.color);
-    }
-}
