--- src/Dz_3_2/ExcelentPupil.java
@@ -1,4 +1,0 @@
-package Dz_3_2;
-
-public class ExcelentPupil extends Pupil{
-}
