--- src/Dz_3_2/GoodPupil.java
@@ -1,5 +1,0 @@
-package Dz_3_2;
-
-public class GoodPupil extends  Pupil{
-
-}
