--- src/Dz_3_2/KlassRoom.java
@@ -1,6 +1,0 @@
-package Dz_3_2;
-
-public class KlassRoom {
-    public KlassRoom() {
-    }
-}
