--- src/Dz_3_2/Pupil.java
@@ -1,16 +1,0 @@
-package Dz_3_2;
-
-public class Pupil {
-
-    public void study (){
-    }
-    public void read(){
-
-    }
-    public void write(){
-
-    }
-    public  void relax(){
-
-    }
-}
