--- src/Dz_3_3/Car.java
@@ -1,10 +1,0 @@
-package Dz_3_3;
-
-public class Car extends Vehicle {
-    public String model;
-
-    public Car(int coordinates, int prices, int speed, int ears, String model) {
-        super(coordinates, prices, speed, ears);
-        this.model = model;
-    }
-}
