--- src/Dz_3_3/Plane.java
@@ -1,13 +1,0 @@
-package Dz_3_3;
-
-public class Plane extends Vehicle{
-    public int height;
-
-    public Plane(int coordinates, int prices, int speed, int ears, int height, int passenger) {
-        super(coordinates, prices, speed, ears);
-        this.height = height;
-        this.passenger = passenger;
-    }
-
-    public int passenger;
-}
