--- src/Dz_3_4/DocumentWorker.java
@@ -1,15 +1,0 @@
-package Dz_3_4;
-
-public class DocumentWorker {
-    public void openDocument(){
-        System.out.println("Документ открыт");
-    }
-
-    public void editDocument(){
-        System.out.println("Редактирование документа доступно в версии Про");
-    }
-
-    public void saveDocument(){
-        System.out.println("Сохранение документа доступно в версии Про");
-    }
-}
